Update to 2.0.3.0
．One vertex can be now control by more than 4 bones, but this kind of mesh will use software skinning only.
