+++ src/main/java/com/modularmods/mcgltf/RenderedGltfModelGL40.java
@@ -60,54 +60,101 @@ public class RenderedGltfModelGL40 extends RenderedGltfModel {
 		ArrayList<Runnable> shaderModNodeRenderCommands = new ArrayList<Runnable>();
 		SkinModel skinModel = nodeModel.getSkinModel();
 		if(skinModel != null) {
-			int jointCount = skinModel.getJoints().size();
-			int jointMatrixSize = jointCount * 16;
+			boolean canHaveHardwareSkinning;
+			checkHardwareSkinning: {
+				for(MeshModel meshModel : nodeModel.getMeshModels()) {
+					for(MeshPrimitiveModel meshPrimitiveModel : meshModel.getMeshPrimitiveModels()) {
+						if(!meshPrimitiveModel.getAttributes().containsKey("JOINTS_1")) {
+							canHaveHardwareSkinning = true;
+							break checkHardwareSkinning;
+						}
+					}
+				}
+				canHaveHardwareSkinning = false;
+			}
 			
-			int jointMatrixBuffer = GL15.glGenBuffers();
-			gltfRenderData.add(() -> GL15.glDeleteBuffers(jointMatrixBuffer));
-			GL15.glBindBuffer(GL31.GL_TEXTURE_BUFFER, jointMatrixBuffer);
-			GL15.glBufferData(GL31.GL_TEXTURE_BUFFER, jointMatrixSize * Float.BYTES, GL15.GL_STATIC_DRAW);
-			int glTexture = GL11.glGenTextures();
-			gltfRenderData.add(() -> GL11.glDeleteTextures(glTexture));
-			GL11.glBindTexture(GL31.GL_TEXTURE_BUFFER, glTexture);
-			GL31.glTexBuffer(GL31.GL_TEXTURE_BUFFER, GL30.GL_RGBA32F, jointMatrixBuffer);
+			int jointCount = skinModel.getJoints().size();
 			
 			float[][] transforms = new float[jointCount][];
 			float[] invertNodeTransform = new float[16];
 			float[] bindShapeMatrix = new float[16];
-			float[] jointMatrices = new float[jointMatrixSize];
 			
-			List<Runnable> jointMatricesTransformCommands = new ArrayList<Runnable>(jointCount);
-			for(int joint = 0; joint < jointCount; joint++) {
-				int i = joint;
-				float[] transform = transforms[i] = new float[16];
-				float[] inverseBindMatrix = new float[16];
-				jointMatricesTransformCommands.add(() -> {
-					MathUtils.mul4x4(invertNodeTransform, transform, transform);
-					skinModel.getInverseBindMatrix(i, inverseBindMatrix);
-					MathUtils.mul4x4(transform, inverseBindMatrix, transform);
-					MathUtils.mul4x4(transform, bindShapeMatrix, transform);
-					System.arraycopy(transform, 0, jointMatrices, i * 16, 16);
-				});
-			}
-			
-			nodeSkinningCommands.add(() -> {
-				for(int i = 0; i < transforms.length; i++) {
-					System.arraycopy(findGlobalTransform(skinModel.getJoints().get(i)), 0, transforms[i], 0, 16);
-				}
-				MathUtils.invert4x4(findGlobalTransform(nodeModel), invertNodeTransform);
-				skinModel.getBindShapeMatrix(bindShapeMatrix);
-				jointMatricesTransformCommands.parallelStream().forEach(Runnable::run);
+			if(canHaveHardwareSkinning) {
+				int jointMatrixSize = jointCount * 16;
+				float[] jointMatrices = new float[jointMatrixSize];
 				
+				int jointMatrixBuffer = GL15.glGenBuffers();
+				gltfRenderData.add(() -> GL15.glDeleteBuffers(jointMatrixBuffer));
 				GL15.glBindBuffer(GL31.GL_TEXTURE_BUFFER, jointMatrixBuffer);
-				GL15.glBufferSubData(GL31.GL_TEXTURE_BUFFER, 0, putFloatBuffer(jointMatrices));
-				
+				GL15.glBufferData(GL31.GL_TEXTURE_BUFFER, jointMatrixSize * Float.BYTES, GL15.GL_STATIC_DRAW);
+				int glTexture = GL11.glGenTextures();
+				gltfRenderData.add(() -> GL11.glDeleteTextures(glTexture));
 				GL11.glBindTexture(GL31.GL_TEXTURE_BUFFER, glTexture);
-			});
-
-			for(MeshModel meshModel : nodeModel.getMeshModels()) {
-				for(MeshPrimitiveModel meshPrimitiveModel : meshModel.getMeshPrimitiveModels()) {
-					processMeshPrimitiveModel(gltfRenderData, nodeModel, meshModel, meshPrimitiveModel, nodeSkinningCommands, vanillaNodeRenderCommands, shaderModNodeRenderCommands);
+				GL31.glTexBuffer(GL31.GL_TEXTURE_BUFFER, GL30.GL_RGBA32F, jointMatrixBuffer);
+				
+				List<Runnable> jointMatricesTransformCommands = new ArrayList<Runnable>(jointCount);
+				for(int joint = 0; joint < jointCount; joint++) {
+					int i = joint;
+					float[] transform = transforms[i] = new float[16];
+					float[] inverseBindMatrix = new float[16];
+					jointMatricesTransformCommands.add(() -> {
+						MathUtils.mul4x4(invertNodeTransform, transform, transform);
+						skinModel.getInverseBindMatrix(i, inverseBindMatrix);
+						MathUtils.mul4x4(transform, inverseBindMatrix, transform);
+						MathUtils.mul4x4(transform, bindShapeMatrix, transform);
+						System.arraycopy(transform, 0, jointMatrices, i * 16, 16);
+					});
+				}
+				
+				nodeSkinningCommands.add(() -> {
+					for(int i = 0; i < transforms.length; i++) {
+						System.arraycopy(findGlobalTransform(skinModel.getJoints().get(i)), 0, transforms[i], 0, 16);
+					}
+					MathUtils.invert4x4(findGlobalTransform(nodeModel), invertNodeTransform);
+					skinModel.getBindShapeMatrix(bindShapeMatrix);
+					jointMatricesTransformCommands.parallelStream().forEach(Runnable::run);
+					
+					GL15.glBindBuffer(GL31.GL_TEXTURE_BUFFER, jointMatrixBuffer);
+					GL15.glBufferSubData(GL31.GL_TEXTURE_BUFFER, 0, putFloatBuffer(jointMatrices));
+					
+					GL11.glBindTexture(GL31.GL_TEXTURE_BUFFER, glTexture);
+				});
+				
+				for(MeshModel meshModel : nodeModel.getMeshModels()) {
+					for(MeshPrimitiveModel meshPrimitiveModel : meshModel.getMeshPrimitiveModels()) {
+						processMeshPrimitiveModel(gltfRenderData, nodeModel, meshModel, meshPrimitiveModel, transforms, nodeSkinningCommands, vanillaNodeRenderCommands, shaderModNodeRenderCommands);
+					}
+				}
+			}
+			else {
+				List<Runnable> jointMatricesTransformCommands = new ArrayList<Runnable>(jointCount);
+				for(int joint = 0; joint < jointCount; joint++) {
+					int i = joint;
+					float[] transform = transforms[i] = new float[16];
+					float[] inverseBindMatrix = new float[16];
+					jointMatricesTransformCommands.add(() -> {
+						MathUtils.mul4x4(invertNodeTransform, transform, transform);
+						skinModel.getInverseBindMatrix(i, inverseBindMatrix);
+						MathUtils.mul4x4(transform, inverseBindMatrix, transform);
+						MathUtils.mul4x4(transform, bindShapeMatrix, transform);
+					});
+				}
+				
+				Runnable jointMatricesTransformCommand = () -> {
+					for(int i = 0; i < transforms.length; i++) {
+						System.arraycopy(findGlobalTransform(skinModel.getJoints().get(i)), 0, transforms[i], 0, 16);
+					}
+					MathUtils.invert4x4(findGlobalTransform(nodeModel), invertNodeTransform);
+					skinModel.getBindShapeMatrix(bindShapeMatrix);
+					jointMatricesTransformCommands.parallelStream().forEach(Runnable::run);
+				};
+				vanillaNodeRenderCommands.add(jointMatricesTransformCommand);
+				shaderModNodeRenderCommands.add(jointMatricesTransformCommand);
+				
+				for(MeshModel meshModel : nodeModel.getMeshModels()) {
+					for(MeshPrimitiveModel meshPrimitiveModel : meshModel.getMeshPrimitiveModels()) {
+						processMeshPrimitiveModel(gltfRenderData, nodeModel, meshModel, meshPrimitiveModel, transforms, vanillaNodeRenderCommands, shaderModNodeRenderCommands);
+					}
 				}
 			}
 		}
