+++ updates.json
@@ -1,12 +1,12 @@
 {
 	"homepage": "https://www.curseforge.com/minecraft/mc-mods/mcgltf/",
 	"1.19.2": {
-		"1.19.2-Forge-2.0.2.1": "Fix morph target generation.",
+		"1.19.2-Forge-2.0.3.0": "Support more JOINTS_n and WEIGHTS_n, and various bug fix.",
 		"1.19.2-Forge-2.0.2.0": "More proper solution to the POM of BSL Shaders.",
 		"1.19.2-Forge-2.0.0.0": "Update to meet Vanilla's minimun OpenGL requirement."
 	},
 	"1.19": {
-		"1.19.2-Forge-2.0.2.1": "Fix morph target generation.",
+		"1.19.2-Forge-2.0.3.0": "Support more JOINTS_n and WEIGHTS_n, and various bug fix.",
 		"1.19.2-Forge-2.0.2.0": "More proper solution to the POM of BSL Shaders.",
 		"1.19.2-Forge-2.0.0.0": "Update to meet Vanilla's minimun OpenGL requirement.",
 		"1.19-Forge-1.2.0.1": "Fix animation cause game crash, and an optimization forgot to reimplement in last update.",
@@ -15,9 +15,9 @@
 		"1.19-Forge-1.0.0.0": "Initial release"
 	},
 	"promos": {
-		"1.19.2-latest": "1.19.2-Forge-2.0.2.1",
-		"1.19.2-recommended": "1.19.2-Forge-2.0.2.1",
-		"1.19-latest": "1.19.2-Forge-2.0.2.1",
-		"1.19-recommended": "1.19.2-Forge-2.0.2.1"
+		"1.19.2-latest": "1.19.2-Forge-2.0.3.0",
+		"1.19.2-recommended": "1.19.2-Forge-2.0.3.0",
+		"1.19-latest": "1.19.2-Forge-2.0.3.0",
+		"1.19-recommended": "1.19.2-Forge-2.0.3.0"
 	}
 }
